add setting to limit max number of colelctions
